General clean up. Updated uniswap tests to typescript
