+++ package.json
@@ -74,6 +74,7 @@
     "prettier-plugin-solidity": "^1.0.0-dev.21",
     "solhint": "^3.3.7",
     "solhint-plugin-prettier": "^0.0.5",
+    "ts-mocha": "^9.0.2",
     "ts-node": "^10.7.0",
     "typechain": "^8.0.0",
     "typescript": "^4.6.3",
